[GENERAL] - Prep work.
Getting the 32-bit build up and running again, complete with process code and able to run CLI etc.
Aiming to get stdlib, and userspace programs (CLI, BASIC, HELLO) all building in 32/64-bit.

Work to do yet, but checking in the current state as it is deploying successfully in full in 32-bit at this point.

+++ c/common/stdio.c
@@ -18,6 +18,7 @@
 #include "stdio.h"
 #include "swis.h"
 #include "kernel.h"
+#include <stdint.h>
 
 
 int sprintf( char *s )
@@ -26,7 +27,7 @@ int sprintf( char *s )
 	struct _kernel_regs in,out;
 	unsigned int return_value;
 
-	in.r[ 0 ] = (unsigned int) s;
+	in.r[ 0 ] = (uintptr_t) s;
 	SWI( OS_PrintString, &in, &out );
 
 	return_value = 0;
